+++ modules/database/src/std/softIoc/softMain.cpp
@@ -17,10 +17,12 @@
 
 #include <epicsGetopt.h>
 #include "registryFunction.h"
+#include "errlog.h"
 #include "epicsThread.h"
 #include "epicsExit.h"
 #include "epicsStdio.h"
 #include "epicsString.h"
+#include "errlog.h"
 #include "dbStaticLib.h"
 #include "subRecord.h"
 #include "dbAccess.h"
@@ -92,7 +94,7 @@ void usage(const char *arg0, const std::string& base_dbd) {
                "interactive IOC shell.\n"
                "\n"
                "Compiled-in path to softIoc.dbd is:\n"
-               "\t"<<base_dbd.c_str()<<"\n";
+               "\t"<<base_dbd.c_str()<<std::endl;
 }
 
 void errIf(int ret, const std::string& msg)
@@ -239,7 +241,9 @@ int main(int argc, char *argv[])
         if (loadedDb) {
             if (verbose)
                 std::cout<<"iocInit()\n";
-            iocInit();
+            if(iocInit()) {
+                std::cerr<<ERL_ERROR " during iocInit()"<<std::endl;
+            }
             epicsThreadSleep(0.2);
         }
 
@@ -270,7 +274,8 @@ int main(int argc, char *argv[])
         return 0;
 
     }catch(std::exception& e){
-        std::cerr<<"Error: "<<e.what()<<"\n";
+        errlogFlush();
+        std::cerr<<ERL_ERROR ": "<<e.what()<<"\n";
         epicsExit(2);
         return 2;
     }